+++ src/sammie/js/ui/elements/MasksOverImage.tsx
@@ -4,8 +4,15 @@ import {useRecoilValue} from "recoil";
 import * as ui from "../../state/uistates";
 
 export type MaskOverImageMask = {
+    /**The url for the mask image to display*/
     url:string,
-    col:'black'|'blue'|'red'|'white',
+    /**Filters to apply to mask. Works only for binary masks*/
+    col:'black'|'blue'|'red'|'white'|'original',
+    /**Opacity of mask*/
+    opacity?:number,
+    /**Shift the mask by an amount of pixels*/
+    shift?:[number,number]|number[],
+    /**Additional class names of the mask image*/
     className?:string
 }
 interface IMasksOverImageProps{
@@ -30,8 +37,14 @@ const MasksOverImage:React.FC<IMasksOverImageProps> = ({ className,originalURL,s
     
             {masks.map((m,i)=>{
                 if(m === null || m === false) return;
+                
+                var maskStyle = {opacity:1,transform:'none'};
+                if(m.opacity !== undefined) maskStyle.opacity = m.opacity;
+                if(m.shift !== undefined) maskStyle.transform = `translate(${m.shift[0]}px,${m.shift[1]}px)`;
+                
                 return <img key={i} src={m.url} alt={'Mask Image'}
-                         className={`mask-img ${m.col} ${m.className}`}/>
+                         className={`mask-img ${m.col} ${m.className}`}
+                            style={maskStyle}/>
             }
             )}
         </div>